Normal Vector In range [-1,1].
